\name{soft.threshold}
\alias{soft.threshold}
\title{Soft-thresholding}
\description{Apply the soft-threshold function to a vector.}
\usage{
soft.threshold(x, lambda = 1)
}
\arguments{
  \item{x}{A vector of numeric data.}
  \item{lambda}{The largest absolute value that will be mapped to zero.}
}

\value{The vector of observations after applying 
the soft-thresholding function.}

\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





