\name{st.inlinedocs-package}
\alias{st.inlinedocs-package}
\docType{package}
\title{Soft-thresholding}
\description{A package documented using inlinedocs.}
\details{
\tabular{ll}{Package: \tab st.inlinedocs\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Soft-thresholding\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


